chore: split overlay effects and messages
