+++ src/main/resources/messages/CodeXPBundle.properties
@@ -25,3 +25,13 @@ TEXT_EFFECT=Effect
 TEXT_TYPE=Type
 settings.notification.type.intellij.description=Default notification will appear in the bottom-right of the IDE and IDE notification tool window.
 settings.notification.type.codexp.description=Customized notification will appear in the top-center of the IDE.
+overlay.level.up.title=Level Up!
+overlay.level.up.main=Congratulations! You are now level {0}!
+overlay.level.up.sub=XP to next level: {0} xp
+overlay.challenge.completed.title=Challenge Completed!
+overlay.challenge.completed.main=Congratulations! You have completed {0}!
+overlay.challenge.completed.sub=XP earned: {0} xp
+notification.challenge.completed.title={0} completed!
+notification.challenge.completed.content=Reward: {0} XP
+notification.level.up.title=Level up!
+notification.level.up.content=Congratulations {0}! You are now level {1}. You need {2} XP to reach the next level.
